Change packet payload to uint32_t.
Also fix ntoh ordering on packet deserialization.

+++ hardware/APS/libaps2-cpp/src/APS/APSEthernetPacket.h
@@ -25,7 +25,7 @@ struct APSEthernetHeader {
 class APSEthernetPacket{
 public:
 	APSEthernetHeader header;
-	vector<uint8_t> payload;
+	vector<uint32_t> payload;
 
 	APSEthernetPacket();
